requirements.txt
# CEIEF Project Dependencies
# Cultural Education Interaction Enhancement Framework

# Core LLM & API
anthropic>=0.25.0
openai>=1.0.0
httpx>=0.24.0
requests>=2.31.0

# Knowledge Graph
neo4j>=5.14.0
networkx>=3.1
rdflib>=7.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# NLP & Text Processing
transformers>=4.35.0
sentence-transformers>=2.2.2
nltk>=3.8.1
jieba>=0.42.1
langdetect>=1.0.9

# Embeddings & Vector Search
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging & Monitoring
loguru>=0.7.2
tqdm>=4.66.0

# Evaluation & Metrics
rouge-score>=0.1.2
nltk>=3.8.1
bert-score>=0.3.13

# Utilities
python-dateutil>=2.8.2
uuid>=1.30
jsonschema>=4.20.0
rich>=13.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3
